Remove Confederate from color list
Just joined an open source project that's using this and [animals](https://github.com/boennemann/animals/blob/master/words.json) to generate some random titles (ex: ExpeditionPorcupine289, BonanzaAlligator021, etc....)

Was confused as hell when I saw 
![image](https://user-images.githubusercontent.com/45471522/218232001-a340bf10-10bf-4d67-9341-15efbf402d92.png)

Will admit that it's an accurate color description for #5c6272, but given that so many people likely use this library for so many different things, having "Confederate" in a widely used list of color titles is likely to have some weird unintended consequences. It'd probably be best to just remove.