feat(colors): add magical girl – lively pink
note from the creator: i play colordle everyday and i saw that THERE WAS NO MAGICAL GIRL colour, so
i thought it'd contribute. i know zilch about coding and made a github account just for this. i
understand if you reject since there's already so many colours, but i think it's such a cute colour
and i'd love to use it as my starting word in colordle!!! thanks for your consideration and making
this lovely list!!! 🙇‍♀️